% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_friendly.R
\name{cb_friendly_cols}
\alias{cb_friendly_cols}
\title{Fetch one or more colours by name}
\usage{
cb_friendly_cols(...)
}
\arguments{
\item{...}{One or more colour names (unquoted or quoted strings).
Use \code{\link[=list_cb_friendly_cols]{list_cb_friendly_cols()}} to see all available names.}
}
\value{
Named character vector of hex colour codes.
}
\description{
Fetch one or more colours by name
}
