% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_friendly.R
\name{list_cb_friendly_cols}
\alias{list_cb_friendly_cols}
\title{List all rcore colorblind-friendly colours}
\usage{
list_cb_friendly_cols()
}
\description{
Returns a named character vector of all available colours and their hex codes.
}
