% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_friendly.R
\name{scale_color_cb_friendly}
\alias{scale_color_cb_friendly}
\title{ggplot2 colour scale using rcore palette}
\usage{
scale_color_cb_friendly(
  palette = "main",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{See \code{\link[=cb_friendly_pal]{cb_friendly_pal()}}.}

\item{discrete}{Logical; discrete (\code{TRUE}, default) or continuous scale?}

\item{reverse}{Logical; reverse colour order? Default \code{FALSE}.}

\item{...}{Passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} or
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn()}}.}
}
\value{
A ggplot2 Scale object.
}
\description{
ggplot2 colour scale using rcore palette
}
\examples{
\donttest{
  library(ggplot2)
  ggplot(mtcars, aes(wt, mpg, colour = factor(cyl))) +
    geom_point() +
    scale_color_cb_friendly()
}

}
