% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_friendly.R
\name{scale_fill_cb_friendly}
\alias{scale_fill_cb_friendly}
\title{ggplot2 fill scale using rcore palette}
\usage{
scale_fill_cb_friendly(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{See \code{\link[=cb_friendly_pal]{cb_friendly_pal()}}.}

\item{discrete}{Logical; discrete (\code{TRUE}, default) or continuous scale?}

\item{reverse}{Logical; reverse colour order? Default \code{FALSE}.}

\item{...}{Passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} or
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradientn()}}.}
}
\description{
ggplot2 fill scale using rcore palette
}
