% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_deps.R
\name{list_rcore_deps}
\alias{list_rcore_deps}
\title{List rcore dependencies by analysis type}
\usage{
list_rcore_deps()
}
\value{
Invisibly returns the dependency list as a named list, where each
element corresponds to an analysis type and contains the required packages.
}
\description{
Prints a summary of all packages required for each analysis type.
}
\examples{
list_rcore_deps()
}
